Move all lucene logic into central builder helpers file
